fix: use reference counter pointer because the lifetime yada yada yada
